# docs/openapi-fhir.yaml
openapi: 3.1.0
info:
  title: PCM FHIR API
  version: "2026-01-05"
  description: |
    OpenAPI definition for the PCM FHIR server. This covers the external FHIR REST API
    interactions described in pcm_process_flow.md, including organization onboarding
    updates, service registration, consent creation and retrieval, approvals, and
    endpoint registration. Examples mirror the FHIR examples under input/fsh/_Examples.
    Unless otherwise noted, FHIR REST requests/responses use application/fhir+json.
servers:
  - url: https://pcm.fhir.health.gov.il/r4
    description: PCM FHIR base URL
...
security:
  - mutualTLS: []
  - bearerAuth: []
paths:
...
  /Organization:
    get:
      tags: [Organization]
      summary: Search organizations
      description: |
        Search organizations by supported parameters. Used for discovery and retrieval,
        including searches by active flag or identifier.
      parameters:
        - $ref: '#/components/parameters/_id'
        - $ref: '#/components/parameters/identifier'
        - $ref: '#/components/parameters/active'
        - $ref: '#/components/parameters/type'
        - $ref: '#/components/parameters/partof'
        - $ref: '#/components/parameters/_include'
        - $ref: '#/components/parameters/_include_iterate'
      responses:
        '200':
          description: Bundle of matching organizations
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Bundle'
              examples:
                organizationSearch:
                  $ref: '#/components/examples/OrganizationSearchBundle'
        '400':
          $ref: '#/components/responses/OperationOutcomeBadRequest'
        '401':
          $ref: '#/components/responses/OperationOutcomeUnauthorized'
        '403':
          $ref: '#/components/responses/OperationOutcomeForbidden'
        default:
          $ref: '#/components/responses/OperationOutcomeError'
...
  /Endpoint:
    get:
      tags: [Endpoint]
      summary: Search endpoints
      parameters:
        - $ref: '#/components/parameters/_id'
        - $ref: '#/components/parameters/thumbprint'
      responses:
        '200':
          description: Bundle of endpoints
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Bundle'
              examples:
                endpointSearch:
                  $ref: '#/components/examples/EndpointSearchBundle'
        '400':
          $ref: '#/components/responses/OperationOutcomeBadRequest'
        '401':
          $ref: '#/components/responses/OperationOutcomeUnauthorized'
        '403':
          $ref: '#/components/responses/OperationOutcomeForbidden'
        default:
          $ref: '#/components/responses/OperationOutcomeError'
    post:
      tags: [Endpoint]
      operationId: createEndpoint
      summary: Create endpoint
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: '#/components/schemas/Endpoint'
            examples:
              endpointSubmit:
                $ref: '#/components/examples/EndpointSubmit'
      responses:
        '201':
          description: Created endpoint
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Endpoint'
              examples:
                endpointCreated:
                  $ref: '#/components/examples/EndpointRetrieval'
          links:
            ReadEndpointById:
              operationId: readEndpoint
              parameters:
                id: '$response.body#/id'
              description: Follow-up read of the created Endpoint.
        '400':
          $ref: '#/components/responses/OperationOutcomeBadRequest'
        '401':
          $ref: '#/components/responses/OperationOutcomeUnauthorized'
        '403':
          $ref: '#/components/responses/OperationOutcomeForbidden'
        default:
          $ref: '#/components/responses/OperationOutcomeError'
...
  /Consent:
    get:
      tags: [Consent]
      operationId: searchConsents
      summary: Search consents
      description: |
        Search consents by supported parameters. Use _include to return actor
        organizations, endpoints, and parent orgs in a single bundle.
      parameters:
        - $ref: '#/components/parameters/_id'
        - $ref: '#/components/parameters/identifier'
        - $ref: '#/components/parameters/status'
        - $ref: '#/components/parameters/patient'
        - $ref: '#/components/parameters/pcm_service'
        - $ref: '#/components/parameters/_include'
        - $ref: '#/components/parameters/_include_iterate'
      responses:
        '200':
          description: Bundle of consents
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Bundle'
              examples:
                consentIncludeBundle:
                  $ref: '#/components/examples/ConsentIncludeBundle'
        '400':
          $ref: '#/components/responses/OperationOutcomeBadRequest'
        '401':
          $ref: '#/components/responses/OperationOutcomeUnauthorized'
        '403':
          $ref: '#/components/responses/OperationOutcomeForbidden'
        default:
          $ref: '#/components/responses/OperationOutcomeError'
    post:
      tags: [Consent]
      operationId: createConsent
      summary: Create consent request
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: '#/components/schemas/Consent'
            examples:
              consentRequest:
                $ref: '#/components/examples/ConsentRequest'
      responses:
        '201':
          description: Created consent
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Consent'
              examples:
                consentCreated:
                  $ref: '#/components/examples/ConsentRequestCreated'
          links:
            ReadConsentById:
              operationId: readConsent
              parameters:
                id: '$response.body#/id'
              description: Follow-up read of the created Consent.
        '400':
          $ref: '#/components/responses/OperationOutcomeBadRequest'
        '401':
          $ref: '#/components/responses/OperationOutcomeUnauthorized'
        '403':
          $ref: '#/components/responses/OperationOutcomeForbidden'
        default:
          $ref: '#/components/responses/OperationOutcomeError'
...
components:
  securitySchemes:
    mutualTLS:
      type: mutualTLS
      description: mTLS client authentication.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: OAuth2 bearer token (opaque) required for PCM FHIR API access.
  parameters:
...
    _id:
      name: _id
      in: query
      required: false
      schema:
        type: string
    identifier:
      name: identifier
      in: query
      required: false
      schema:
        type: string
      description: FHIR identifier search parameter.
    thumbprint:
      name: thumbprint
      in: query
      required: false
      schema:
        type: string
      description: Search by NTN-registered applicable certificate thumbprint (x5t#S256, base64url SHA-256) stored in the Endpoint extension.
    active:
      name: active
      in: query
      required: false
      schema:
        type: string
      description: Search by activation status (token; true|false).
...
    status:
      name: status
      in: query
      required: false
      schema:
        type: string
    type:
      name: type
      in: query
      required: false
      schema:
        type: string
      description: Search by type code (Organization.type or HealthcareService.type).
...
    partof:
      name: partof
      in: query
      required: false
      schema:
        type: string
    patient:
      name: patient
      in: query
      required: false
      schema:
        type: string
      description: Search by patient logical reference; supports patient.identifier chain (system|value).
    pcm_service:
      name: pcm-service
      in: query
      required: false
      schema:
        type: string
      description: Search by pcmService extension reference (HealthcareService instance).
    _include:
      name: _include
      in: query
      required: false
      schema:
        type: string
      description: Include related resources (FHIR _include).
    _include_iterate:
      name: _include:iterate
      in: query
      required: false
      schema:
        type: string
      description: Recursive include for related resources.
  responses:
    OperationOutcomeBadRequest:
      description: Bad request (OperationOutcome)
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
          examples:
            badRequest:
              $ref: '#/components/examples/OperationOutcomeBadRequest'
    OperationOutcomeUnauthorized:
      description: Unauthorized (OperationOutcome)
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
          examples:
            unauthorized:
              $ref: '#/components/examples/OperationOutcomeUnauthorized'
    OperationOutcomeForbidden:
      description: Forbidden (OperationOutcome)
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
          examples:
            forbidden:
              $ref: '#/components/examples/OperationOutcomeForbidden'
...
    OperationOutcomeError:
      description: Server error (OperationOutcome)
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
          examples:
            internalError:
              $ref: '#/components/examples/OperationOutcomeInternalError'
  schemas:
    BaseResource:
      type: object
      required: [resourceType]
      properties:
        resourceType:
          type: string
        id:
          type: string
        meta:
          $ref: '#/components/schemas/Meta'
...
    Bundle:
      allOf:
        - $ref: '#/components/schemas/BaseResource'
        - type: object
          required: [resourceType, type]
          properties:
            resourceType:
              type: string
              enum: [Bundle]
            type:
              type: string
              enum:
                - document
                - message
                - transaction
                - transaction-response
                - batch
                - batch-response
                - history
                - searchset
                - collection
            total:
              type: integer
            link:
              type: array
              items:
                $ref: '#/components/schemas/BundleLink'
            entry:
              type: array
              items:
                $ref: '#/components/schemas/BundleEntry'
    BundleLink:
      type: object
      properties:
        relation:
          type: string
        url:
          type: string
    BundleEntry:
      type: object
      properties:
        resource:
          $ref: '#/components/schemas/Resource'
        search:
          type: object
          properties:
            mode:
              type: string
...
    Endpoint:
      allOf:
        - $ref: '#/components/schemas/DomainResource'
        - type: object
          required: [resourceType]
          properties:
            resourceType:
              type: string
              enum: [Endpoint]
            status:
              type: string
              enum:
                - active
                - suspended
                - error
                - off
                - entered-in-error
                - test
            connectionType:
              $ref: '#/components/schemas/Coding'
            payloadType:
              type: array
              items:
                $ref: '#/components/schemas/CodeableConcept'
            address:
              type: string
            managingOrganization:
              $ref: '#/components/schemas/Reference'
...
    Consent:
      allOf:
        - $ref: '#/components/schemas/DomainResource'
        - type: object
          required: [resourceType]
          properties:
            resourceType:
              type: string
              enum: [Consent]
            status:
              type: string
              enum:
                - proposed
                - active
                - rejected
                - inactive
            dateTime:
              type: string
              format: date-time
            identifier:
              type: array
              items:
                $ref: '#/components/schemas/Identifier'
            scope:
              $ref: '#/components/schemas/CodeableConcept'
            category:
              type: array
              items:
                $ref: '#/components/schemas/CodeableConcept'
            patient:
              $ref: '#/components/schemas/Reference'
            provision:
              $ref: '#/components/schemas/ConsentProvision'
...
    Coding:
      type: object
      properties:
        system:
          type: string
        code:
          type: string
        display:
          type: string
...
  examples:
...
    EndpointSubmit:
      summary: HMO B endpoint (submit)
      value:
        resourceType: Endpoint
        status: active
        connectionType:
          system: "http://terminology.hl7.org/CodeSystem/endpoint-connection-type"
          code: hl7-fhir-rest
          display: "HL7 FHIR REST"
        payloadType:
          - coding:
              - system: "urn:oid:1.3.6.1.4.1.19376.1.2.3"
                code: "urn:ihe:pcc:xphr:2007"
                display: "XPHR Extract"
        address: "https://fhir.hmo-b.example.org/r4"
        managingOrganization:
          reference: "Organization/org-ds-hmo-b-submit"
        extension:
          - url: "http://pcm.fhir.health.gov.il/StructureDefinition/ext-applicable-certificates"
            extension:
              - url: "thumbprint"
                valueString: "dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk"
    EndpointRetrieval:
      summary: HMO B endpoint (retrieval)
      value:
        resourceType: Endpoint
        id: endpoint-hmo-b
        status: active
        connectionType:
          system: "http://terminology.hl7.org/CodeSystem/endpoint-connection-type"
          code: hl7-fhir-rest
          display: "HL7 FHIR REST"
        payloadType:
          - coding:
              - system: "urn:oid:1.3.6.1.4.1.19376.1.2.3"
                code: "urn:ihe:pcc:xphr:2007"
                display: "XPHR Extract"
        address: "https://fhir.hmo-b.example.org/r4"
        managingOrganization:
          reference: "Organization/org-ds-hmo-b"
        extension:
          - url: "http://pcm.fhir.health.gov.il/StructureDefinition/ext-applicable-certificates"
            extension:
              - url: "thumbprint"
                valueString: "dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk"
...
    ConsentRequest:
      summary: Consent request (submit)
      value:
        resourceType: Consent
        status: proposed
        scope:
          coding:
            - system: "http://terminology.hl7.org/CodeSystem/consentscope"
              code: patient-privacy
              display: "Patient Privacy"
        category:
          - coding:
              - system: "http://terminology.hl7.org/CodeSystem/v3-ActCode"
                code: INFA
                display: "information access"
        patient:
          identifier:
            system: "http://fhir.health.gov.il/identifier/il-national-id"
            value: "324573740"
        extension:
          - url: "http://pcm.fhir.health.gov.il/StructureDefinition/ext-pcm-service"
            valueReference:
              reference: "HealthcareService/svc-instance-hosp-a-cross-org-summary"
        provision:
          type: permit
          purpose:
            - system: "http://terminology.hl7.org/CodeSystem/v3-ActReason"
              code: TREAT
              display: "treatment"
    ConsentRequestCreated:
      summary: Consent request (created)
      value:
        resourceType: Consent
        id: consent-request-created-hosp-a-cross-org-summary
        meta:
          lastUpdated: "2026-01-05T08:10:00Z"
        identifier:
          - use: official
            system: "http://pcm.fhir.health.gov.il/identifier/pcm-consent-id"
            value: "PCM-CONSENT-0101"
        status: proposed
        dateTime: "2026-01-05T10:10:00+02:00"
        scope:
          coding:
            - system: "http://terminology.hl7.org/CodeSystem/consentscope"
              code: patient-privacy
              display: "Patient Privacy"
        category:
          - coding:
              - system: "http://fhir.health.gov.il/cs/pcm-consent-access-mode"
                code: continuous
                display: "Continuous"
          - coding:
              - system: "http://terminology.hl7.org/CodeSystem/v3-ActCode"
                code: INFA
                display: "information access"
        patient:
          identifier:
            system: "http://fhir.health.gov.il/identifier/il-national-id"
            value: "324573740"
        extension:
          - url: "http://pcm.fhir.health.gov.il/StructureDefinition/ext-pcm-service"
            valueReference:
              reference: "HealthcareService/svc-instance-hosp-a-cross-org-summary"
        provision:
          type: permit
          purpose:
            - system: "http://terminology.hl7.org/CodeSystem/v3-ActReason"
              code: TREAT
              display: "treatment"
          actor:
            - role:
                coding:
                  - system: "http://terminology.hl7.org/CodeSystem/v3-ParticipationType"
                    code: IRCP
                    display: "information recipient"
              reference:
                reference: "Organization/org-sp-hospital-a"
...
    OrganizationSearchBundle:
      summary: Organization search bundle
      value:
        resourceType: Bundle
        id: bundle-organization-search
        type: searchset
        total: 2
        entry:
          - search:
              mode: match
            resource:
              resourceType: Organization
              id: org-parent-hospital-a
              name: "Hospital A (Parent Legal Entity)"
          - search:
              mode: match
            resource:
              resourceType: Organization
              id: org-parent-hmo-b-submit
              name: "HMO B (Parent Legal Entity)"
    EndpointSearchBundle:
      summary: Endpoint search bundle
      value:
        resourceType: Bundle
        id: bundle-endpoint-search
        type: searchset
        total: 1
        entry:
          - search:
              mode: match
            resource:
              resourceType: Endpoint
              id: endpoint-hmo-b
              address: "https://fhir.hmo-b.example.org/r4"
...